+++ src/minfraud/validation.py
@@ -285,7 +285,9 @@ _event_type = In(
     [
         "account_creation",
         "account_login",
+        "credit_application",
         "email_change",
+        "fund_transfer",
         "password_reset",
         "payout_change",
         "purchase",