+++ tests/test_validation.py
@@ -324,7 +324,9 @@ class TestEvent(ValidationBase, unittest.TestCase):
         for good in (
             "account_creation",
             "account_login",
+            "credit_application",
             "email_change",
+            "fund_transfer",
             "password_reset",
             "payout_change",
             "purchase",