+++ HISTORY.rst
@@ -9,6 +9,8 @@ History
 * Setuptools has been replaced with the uv build backend for building the
   package.
 * Added ``securepay`` to the ``/payment/processor`` validation.
+* Added ``credit_application`` and ``fund_transfer`` to the ``/event/type``
+  validation.
 
 3.1.0 (2025-05-23)
 ++++++++++++++++++