+++ HISTORY.rst
@@ -6,6 +6,8 @@ History
 3.2.0
 ++++++++++++++++++
 
+* IMPORTANT: Python 3.10 or greater is required. If you are using an older
+  version, please use an earlier release.
 * Setuptools has been replaced with the uv build backend for building the
   package.
 * Added ``securepay`` to the ``/payment/processor`` validation.