+++ HISTORY.rst
@@ -11,8 +11,8 @@ History
 * Setuptools has been replaced with the uv build backend for building the
   package.
 * Added ``securepay`` to the ``/payment/processor`` validation.
-* Added ``credit_application`` and ``fund_transfer`` to the ``/event/type``
-  validation.
+* Added ``credit_application``, ``fund_transfer``, and ``sim_swap`` to the
+  ``/event/type`` validation.
 * Added the input ``/event/party``. This is the party submitting the
   transaction.
 * Added the input ``/payment/method``. This is the payment method associated