+++ HISTORY.rst
@@ -3,7 +3,7 @@
 History
 -------
 
-3.2.0
+3.2.0 (2025-11-20)
 ++++++++++++++++++
 
 * IMPORTANT: Python 3.10 or greater is required. If you are using an older